requirements.txt
pydantic
alembic
psycopg2-binary
requests
uvicorn
boto3
fastapi
botocore
SQLAlchemy